Resolve problems with emits from UpdateTaskComponent to Dasboard
