+++ src/views/Auth.vue
@@ -1,15 +1,12 @@
 <script setup>
 import { ref } from 'vue'
-import SingUpComponent from '../components/SingUpComponent.vue'
-import LogInComponent from '../components/LogInComponent.vue'
-
+import SingUpComponent from '../components/user/SingUpComponent.vue'
+import LogInComponent from '../components/user/LogInComponent.vue'
 
 const hasUser = ref(true)
 function toggleUser() {
   hasUser.value = !hasUser.value
 }
-
-
 </script>
 
 <template>
@@ -18,37 +15,35 @@ function toggleUser() {
     <article v-if="hasUser">
       <LogInComponent />
       <div class="change-component">
-      <p>Don't have an account?</p>
-      <a href="#" @click="toggleUser">Create</a>
-    </div>
+        <p>Don't have an account?</p>
+        <a href="#" @click="toggleUser">Create</a>
+      </div>
     </article>
 
     <article v-else="!hasUser">
       <SingUpComponent />
       <div class="change-component">
-      <p>Already have an account?</p>
-      <a href="#" @click="toggleUser">Log In</a>
-    </div>
-
+        <p>Already have an account?</p>
+        <a href="#" @click="toggleUser">Log In</a>
+      </div>
     </article>
   </section>
 </template>
 
 <style scoped>
-.auth{
+.auth {
   padding-top: 60px;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   gap: 60px;
-  
 }
-h1{
+h1 {
   color: var(--purple);
   font-size: 35px;
 }
-.change-component{
+.change-component {
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -56,16 +51,16 @@ h1{
   padding-top: 10px;
   gap: 10px;
 }
-p{
+p {
   color: var(--gray);
   font-size: 13px;
 }
-a{
+a {
   color: var(--purple);
   text-decoration: none;
   font-size: 13px;
 }
-a:hover{
+a:hover {
   color: var(--yellow);
 }
 </style>
